requirements.txt
requests==2.30.0
discord.py==2.1.1
discord_webhook
aiofiles==23.1.0
aiosqlite3==0.3.0